# config/swagger.yml
swagger: "2.0"
info:
  version: "1.0.0"
  title: wellness-nudge
  description: |
    On-device wellness nudge agent. Given a short JSON summary of a user's
    recent biometric data (sleep, HRV, resting heart rate, activity), this
    mim returns a short, plain-language nudge generated by a local LLM via
    the mimik AI Foundation inference API. Runs entirely on-device — no
    cloud calls and no personally-identifying data leaves the node.

    Each nudge is classified into a wellness-goal category and stored locally.
    When a new request arrives, past nudges the user marked as helpful for the
    same category may be referenced as a "by the way" subtitle.
  contact:
    name: mimik
    url: "https://mimik.com/"
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: /wellness-nudge/v1
consumes:
  - application/json
produces:
  - application/json
responses:
  "400":
    description: Bad input parameter
    schema:
      $ref: "#/definitions/BadRequest"
  "403":
    description: Forbidden request
    schema:
      $ref: "#/definitions/Forbidden"
  "404":
    description: Not found
    schema:
      $ref: "#/definitions/Error"
  "500":
    description: System error
    schema:
      $ref: "#/definitions/ServerError"
  "502":
    description: Upstream inference error
    schema:
      $ref: "#/definitions/ServerError"
securityDefinitions:
  bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "Bearer token authentication. Use format: Bearer <api-key>"
paths:
  /healthcheck:
    get:
      summary: Liveness probe
      description: Returns 200 if the mim instance is up and responding.
      x-swagger-router-controller: healthController
      operationId: getHealth
      responses:
        200:
          description: ok
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  status: { type: string, example: ok }
                  mim: { type: string, example: wellness-nudge }
                  version: { type: string, example: 1.0.0 }
  /nudge:
    post:
      summary: Generate a wellness nudge
      security:
        - bearer: []
      description: |
        Given recent biometric metrics, returns a short personalized nudge
        in 1 to 3 sentences. The model is prompted to avoid medical claims
        and to ground its suggestion in at least one of the supplied signals.

        The mim also classifies the request into one of nine wellness-goal
        categories and persists the result in local storage. If a past nudge
        from the same category was marked helpful, it may be referenced as
        a "byTheWay" subtitle in the response.
      x-swagger-router-controller: nudgeController
      operationId: createNudge
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/NudgeRequest"
      responses:
        200:
          description: Successful response with the generated nudge.
          schema:
            type: object
            properties:
              data:
                $ref: "#/definitions/NudgeResponse"
        400:
          $ref: "#/responses/400"
        403:
          $ref: "#/responses/403"
        500:
          $ref: "#/responses/500"
        502:
          $ref: "#/responses/502"
  /history:
    get:
      summary: List recent nudges
      security:
        - bearer: []
      description: |
        Returns the most recent nudges this user has generated, newest first.
        Each entry includes the request metrics, generated nudge text, inferred
        category, and the user's feedback status (helpful=yes|no|unset).
      x-swagger-router-controller: historyController
      operationId: listHistory
      parameters:
        - name: limit
          in: query
          type: integer
          required: false
          default: 20
          description: Maximum number of nudges to return (1-100).
      responses:
        200:
          description: List of past nudges.
          schema:
            type: object
            properties:
              data:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/definitions/NudgeHistoryItem"
                  total:
                    type: integer
                    description: Total number of nudges stored.
        403:
          $ref: "#/responses/403"
        500:
          $ref: "#/responses/500"
  /tips:
    get:
      summary: Personal tips based on recent focus and helpful feedback
      security:
        - bearer: []
      description: |
        Aggregates the user's recent nudge categories and surfaces past nudges
        they marked as helpful within those same wellness goals. Read-only and
        cheap to call — no inference, just storage scans. The Android app
        invokes this when the Personal Tips tab is opened or pulled to refresh.

        A "tip card" is emitted per focus category that the user has both
        (a) asked about in the recent window and (b) previously rated helpful.
        Categories with no helpful past nudges are omitted.
      x-swagger-router-controller: tipsController
      operationId: getTips
      responses:
        200:
          description: Tip cards, possibly empty.
          schema:
            type: object
            properties:
              data:
                $ref: "#/definitions/TipsResponse"
        403:
          $ref: "#/responses/403"
        500:
          $ref: "#/responses/500"
  /nudges/{id}/feedback:
    put:
      summary: Mark a past nudge as helpful, unhelpful, or unset
      security:
        - bearer: []
      description: |
        Updates the feedback flag for a past nudge. Nudges marked helpful in a
        given category will surface in that category's Personal Tips card.
      x-swagger-router-controller: historyController
      operationId: updateFeedback
      parameters:
        - name: id
          in: path
          type: string
          required: true
          description: Nudge id, e.g. "nudge_1747700000000".
        - name: body
          in: body
          required: true
          schema:
            $ref: "#/definitions/FeedbackRequest"
      responses:
        200:
          description: Updated nudge entry.
          schema:
            type: object
            properties:
              data:
                $ref: "#/definitions/NudgeHistoryItem"
        400:
          $ref: "#/responses/400"
        403:
          $ref: "#/responses/403"
        404:
          $ref: "#/responses/404"
        500:
          $ref: "#/responses/500"
  /nudges/{id}/clear:
    # POST instead of DELETE: when bundled, mw.delete(...) silently fails to
    # register on the Duktape runtime — likely the reserved-word property
    # access is being mangled. Path makes the intent unambiguous.
    post:
      summary: Permanently remove a past nudge
      security:
        - bearer: []
      description: |
        Removes a nudge from local storage. It will no longer appear in
        /history, won't be referenced by /tips, and any feedback recorded
        against it is lost. Used by the Clear button on the nudge result
        screen.
      x-swagger-router-controller: historyController
      operationId: deleteNudge
      parameters:
        - name: id
          in: path
          type: string
          required: true
          description: Nudge id, e.g. "nudge_1747700000000".
      responses:
        204:
          description: Nudge removed.
        403:
          $ref: "#/responses/403"
        404:
          $ref: "#/responses/404"
        500:
          $ref: "#/responses/500"
definitions:
  Error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        description: HTTP status code mirror.
      message:
        type: string
        description: Human-readable error message.
  BadRequest:
    type: object
    allOf:
      - $ref: "#/definitions/Error"
    example:
      code: 400
      message: Bad input parameter
  Forbidden:
    type: object
    allOf:
      - $ref: "#/definitions/Error"
    example:
      code: 403
      message: forbidden
  ServerError:
    type: object
    allOf:
      - $ref: "#/definitions/Error"
    example:
      code: 500
      message: System error
  NudgeRequest:
    type: object
    description: |
      Recent biometric metrics describing the user. All fields are optional,
      but at least one must be supplied. The mim does not fabricate missing values.
    properties:
      sleepHours:
        type: number
        format: float
        description: Total sleep duration last night, in hours.
        example: 6.2
      deepSleepPct:
        type: number
        format: float
        description: Percentage of sleep classified as deep sleep (0–100).
        example: 14
      remSleepPct:
        type: number
        format: float
        description: Percentage of sleep classified as REM (0–100).
        example: 18
      restingHR:
        type: integer
        description: Resting heart rate in BPM.
        example: 64
      hrvMs:
        type: integer
        description: Heart rate variability (RMSSD-like) in ms.
        example: 42
      stepsYesterday:
        type: integer
        description: Step count from the previous day.
        example: 6800
      userGoal:
        type: string
        description: Short free-text user focus or goal for the day.
        example: improve recovery
  NudgeResponse:
    type: object
    properties:
      id:
        type: string
        description: Unique id of this nudge. Use with PATCH /nudges/{id}/feedback.
        example: "nudge_1747700000000"
      nudge:
        type: string
        description: The generated wellness nudge (1–3 sentences).
        example: Your HRV is on the low side and you got under 6h of sleep. Try a 10-minute walk after lunch and aim to be in bed 30 minutes earlier tonight.
      category:
        type: string
        description: Wellness goal category the request was classified into.
        example: improve-sleep
      model:
        type: string
        description: ID of the local model used for the main nudge.
        example: smollm2-360m
      finishReason:
        type: string
        description: Reason the model stopped generating (e.g. "stop", "length").
        example: stop
      usage:
        type: object
        description: Token usage from the local inference call.
        properties:
          prompt_tokens: { type: integer }
          completion_tokens: { type: integer }
          total_tokens: { type: integer }
      metricsUsed:
        type: object
        description: The sanitized subset of the request body actually sent to the model.
  NudgeHistoryItem:
    type: object
    description: A single past nudge record.
    properties:
      id:
        type: string
        example: "nudge_1747700000000"
      ts:
        type: integer
        format: int64
        description: Unix milliseconds timestamp when the nudge was generated.
        example: 1747700000000
      metrics:
        type: object
        description: The metrics sent with the original request.
      userGoal:
        type: string
        example: improve recovery
      category:
        type: string
        example: improve-recovery
      nudge:
        type: string
      model:
        type: string
        example: smollm2-360m
      helpful:
        type: string
        enum: [yes, "no", unset]
        description: User feedback flag.
        example: unset
  FeedbackRequest:
    type: object
    required:
      - helpful
    properties:
      helpful:
        type: string
        enum: [yes, "no", unset]
        description: New feedback value.
  TipsResponse:
    type: object
    description: Tip cards summarizing the user's recent focus and what helped before.
    properties:
      items:
        type: array
        items:
          $ref: "#/definitions/TipCard"
      totalRecent:
        type: integer
        description: Number of recent nudges considered when building this list.
        example: 14
      generatedAt:
        type: integer
        format: int64
        description: Unix milliseconds — when this tip list was assembled.
        example: 1747700000000
  TipCard:
    type: object
    description: A category-scoped tip with past helpful nudges.
    properties:
      category:
        type: string
        description: Wellness goal category slug.
        example: improve-sleep
      categoryLabel:
        type: string
        description: Friendly label for the category, used in card copy.
        example: improving sleep
      intro:
        type: string
        description: Card subtitle, e.g. "Your recent nudges suggest you've been focused on improving sleep…"
      recentMentions:
        type: integer
        description: How many times this category appeared in the recent window.
        example: 3
      lastFocusedAt:
        type: integer
        format: int64
        description: Unix milliseconds — last time the user generated a nudge in this category.
      helpfulNudges:
        type: array
        items:
          $ref: "#/definitions/TipNudge"
  TipNudge:
    type: object
    properties:
      id:
        type: string
        example: "nudge_1747700000000"
      ts:
        type: integer
        format: int64
        example: 1747700000000
      nudge:
        type: string